Fix the release workflow
The release workflow should not attempt to do an actual release
unless there is a release tag that is pushed.

Also, change some of the push events in workflows to pull_request.
We probably don't need to run these workflows unless we are
actively working towards contributing them to the main branch.